WebUi_Auto/requirements.txt
PyYAML == 3.12
selenium == 3.11.0
logger == 1.1